requirements.txt
# Core pipeline (Phases 1-4)
rich>=13.0.0
faster-whisper>=1.0.0
openai-whisper==20250625       # fallback if faster-whisper unavailable
openai==2.30.0
pydub==0.25.1
python-dotenv==1.2.2

# Whisper runtime deps — pin these exactly (numpy<2 required for torch 2.2.x)
torch==2.2.2
numpy==1.26.4
numba==0.60.0
llvmlite==0.43.0

# Phase 5: Speaker diarization (requires HF_TOKEN in .env)
pyannote.audio>=3.1.0

# Phase 2: Silent Speaker Learning (cosine similarity for voice matching)
scikit-learn>=1.3.0

# Phase 8: TTS (deferred — install when ready)
# TTS>=0.22.0
